requirements.txt
pdf2image>=1.16.3
Pillow>=10.0.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
notebook>=7.0.0
kraken>=3.0.5
numpy>=1.24.0
scikit-learn>=1.3.0